assessment/requirements.txt
flask>=3.0
flask-wtf>=1.2
flask-limiter>=3.5
python-dotenv
pydantic-settings>=2.0
numpy>=1.24
scikit-learn>=1.3
tiktoken>=0.5
gunicorn>=21.0
google-cloud-firestore>=2.16
